Add det(...) and *-broadcasting for polynomial matrices
The generic implementation of det(...) seems to assume the coefficients
are in a field, which is why I'm adding an 'even more generic' version
here. It's the slowest possible implementation and should probably be
replaced by some version of row and column operations.

As for *-broadcasting, this is not happening because Polynomial is not
a Number.