Make syzygies and groebner_basis work for modules
The trickiest thing to debug was that because NamedPolynomial
is not a Number, we get that

    x * [x,y,z]'

works and returns a 1x3 array. However, we need it to return
a RowVector, which is subtly different from an 1x3 array: a RowVector
times a Vector results in a number (not a 1x1 array).

The easiest way to see this issue with built-in types is:

    julia> import Base: transpose, *;
    julia> *(a::String, b::String) = "$a*$b";
    julia> transpose(a::String) = a;

    julia> broadcast(*, 1, [1,2]')
    1×2 RowVector{Int64,Array{Int64,1}}:
     1  2
    julia>  broadcast(*, "1", ["1","2"]')
    1×2 Array{String,2}:
     "1 x 1"  "1 x 2"

Instead of making Polynomial <: Number (which would make it much
harder to deal with coercions from Number -> Polynomial if we also
want to allow polynomial-valued coefficients), we just implement
multiplication by a row vector ourselves.