groebner_basis(): when not requesting a transformation, use sorting to our advantage
In the red(S, result) step, there's roughly two approaches:

    1. for every leading monomial in result, see if it divides lm(S)
    2. for every divisor of lm(S), see if it is in the leading monomials
       in result

It is not obvious which is faster in general; however, in my current
use-case, 2. dominates. That does mean I'll need to keep `result`
sorted, which is a hassle if we want a transformation as well.

I could probably figure it out, but that's a bit useless since when
computing the transformation, the total computation time is easily
dominated by that computation, and not by the reduce step. So I choose
the easy way out and only optimize red(...) using `sort_order` when it
is the bottleneck.