Push typing for variable names down to Monomials
The wrapper type `NamedPolynomial` became more and more cumbersome.

My original idea was to to avoid a combinatorial explosion of code if
people create throw-away variable names; i.e. something like

    julia <<JULIA
    using PolynomialRings
    while true
        R,(x,) = polynomial_ring(gensym())
        x*x; x+x; x-x; x^2 # compile all this code
    end # R is gone, but the code is not
    JULIA

I kind of mis-judged how much effort it would be to maintain this
wrapper; I thought I'd only have to bother for the elementary
operations. I was already proved wrong with the Groebner basis
functions.

If this turns out to be an issue down the line, I'll try to find some
easy wrappers that automatically implement <whatever function of
Polynomials with Named{...}> in terms of <the same function with
Numbered{...}>. I don't anticipate this will happen soon.