Add array versions of all expansion functions/macros
This _joint_iteration thing seemed like a good idea when I started
writing it; I felt it could unify many of these functions. As it turns
out, it unifies only two; and also it doesn't actually work with
Channels as iterations because those are uncomfortable with holding
onto a state for longer than one loop. So looking back, it's a bit
over-engineered now. Oh well.