buchberger(): refactor for understandability, and always return a reduced basis
It requires some bookkeeping to know which pairs of polynomials are
still to be considered, while also changing (reducing, possibly to zero)
the polynomials. This commit moves those manipulations to a few
dedicated functions at the top of the file.

This unified approach makes it rather easy to do the necessary mutual
reductions whenever something changes to the basis. This seems as if
it's more work, but I'm hoping it actually removes some work in common
cases, as it may find that it can remove older polynomials if they are
linear combinations of newer ones.

This removes all use of the dedicated _leadred_grb functions as it is
currently not guaranteed that the result is sorted. I intend to put that
back some time soon.

As it turns out, Julia needs to do some type inference at runtime when
functions in local scope call eachother. To avoid that, I'll have to
refactor to pass the state objects instead of just capturing them, and
define the functions in global scope. I'll do that after I've fully
completed this work, including putting back the sorting optimization.