Display zero polynomial as '0'
This used to be displayed as whatever zero(basering(...)) gives, e.g.
"0//1". This gives a lot of visual clutter for no gain.

The idea of displaying Rational{Int}(0) as "0//1" is to ensure that when
copy-pasting the representation in Julia, one ends up with an object of
the right type. That's already not true for `Rational{BigInt}`, and
it is certainly not true for the polynomial types this commit is about.
So let's not bother with trying to get that right.