Add a version of lift() that computes serveral lifts at once
Computing the Gröbner transformation is the expensive step; after that,
you might as well do a few reductions in a row.

This uses a Tuple but I'm not opposed to adding a version that takes an
AbstractArray. That needs a more complicated type signature to
distinguish the case of several polynomials from the case of a single
module element, and I didn't want to bother.