NumberFields: fix segmentation fault
This seems to have had something to do with using the
@ polynomial macro inside the _ideal function. This caused
the polynomial to be created at macro expansion time
and inserted, as a constant, into the expression tree.

I don't know _why_ it whould cause that, but it seemed
like a scary thing to do and removing it fixes the issue.
So I'll leave it at that.

This doesn't fix all issues with the tests for number fields,
so a few remain commented-out.