Gröbner: Add implicit computation of transformation
Instead of keeping track of the transformation manually (like in
`buchberger`), this commit replaces -- when a transformation is
requested -- the array of polynomials with an array of
`TransformedModuleElement` objects. These behave in all ways
as polynomials except that, under P-linear transformations,
they keep track of the transformation that is being applied.

The code added to `Modules` is still very bare bones and would
need a cleanup.