Fix type inference for a certain conversion
I ran into this while comparing (==) a sparse vector over ZZ[x] with
a sparse vector over QQ[x]. The resulting call to promote would call
this method and would make the array that is now called terms
an Array{Any}.

(I was under the impression the Julia 1.0 doesn't use type inference
for these kinds of things anymore? So I don't quite understand the
issue that this commit solves.)

As a drive-by fix, I noticed the issue with positive characteristic,
so I'm fixing that too. I think I have a few other places where
I have special cases for this and I should add unit tests for
positive characteristic. I can't do that until I have e.g. GaloisFields
as a dependency or as part of this module.