fix xrem(f, g) and friends
The two invocations

    xdiv(f, [g])
    xdiv(f, g)

should basically yield the same result, except that the second form
returns a Polynomial, not a row vector of polynomials. This work with
type specialization of `_unpack` is a hack to save a few lines of code
when unpacking these.

I forgot to update this when I added the in-place versions and the
`x` versions.