Add minimal package requirements to REQUIRE
I was under the impression that this was fully superseded by
Project.toml and Manifest.toml, but that doesn't seem to be
the case:

https://github.com/JuliaLang/METADATA.jl/pull/18494#issuecomment-426413792

It also makes sense as I now realize that nowhere else do we
record _minimal_ versions for dependencies.

I chose the current versions I have installed as it is safe to
assume we need very recent versions to work with 0.7/1.0 at all.