More advanced type promotions
I'd like to take better care of cases such as promoting Int[x]
with Int[x][y][z]. This commit introduces sane (but elaborate)
logic for these cases.

The function promote_rule checks whether the RHS argument's
variables are an ordered subset of the LHS argument's variables.
If so, we can use the LHS, and no complicated promotion is
necessary. The function promote_type calls promote_rule
in both orders.

If neither is a subset of the other, we introduce a 'canonical'
representation of the isomorphism class: it is the one
with all Named variables at the end, and al Numbered
variables in alphabetic order. Example:

    @ring Int[a[]][b[]][c[]][d,e,f]

Some additional logic is needed also to support number fields
and that is part of this commit as well.