Type-inference-frindly integral_fraction
These are two changes that I needed to make to get some
type inference in GWV to work properly. I can't add a
test case because the issue is incredibly intermittent:
any Revise fixes it; who calls it affects it, etc. I'll
just commit these changes and hope for the best.